src/main/java/com/epam/mentoring/patterns/creational/prototype/Employee.java
package com.epam.mentoring.patterns.creational.prototype;

public class Employee implements Cloneable {
	private String name;
	private String position;
	private int salary;

	@Override
	public Employee clone() throws CloneNotSupportedException {
		return (Employee) super.clone();
	}

	public String getName() {
		return name;
	}

	public void setName(String name) {
		this.name = name;
	}

	public String getPosition() {
		return position;
	}

	public void setPosition(String position) {
		this.position = position;
	}

	public int getSalary() {
		return salary;
	}

	public void setSalary(int salary) {
		this.salary = salary;
	}
}
